Adapt create_signatures.sh to create fixtures for Dilithium signatures

+++ spec/fixtures/signatures/create-signatures.sh
@@ -9,16 +9,37 @@ if [ ! -f "$message" ]; then
     dd if=/dev/urandom count=1 bs=64 | base64 > $message
 fi
 
+generate_dilithium_signature=0
+read -p "Generate Dilithium signatures with the OQS-OpenSSH fork? [yN]" -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	generate_dilithium_signatures=1
+	read -p "Path to OQS-OpenSSH root directory: " oqs_openssh_path
+	oqs_ssh_keygen=$oqs_openssh_path"/ssh-keygen"
+	function ssh-keygen-oqs {
+	  "$oqs_ssh_keygen" "$@"
+	}
+fi
+
 create_key_and_sign() {
     local alg=$1
     local keysize=$2
     local key=$filedir/$alg-$keysize-no-options-individual.key
 
     if [ ! -f "$key" ]; then
+      if [ "$alg" == "ssh-dilithium5" ]; then
+        ssh-keygen-oqs -q -N "" -t $alg -C "" -f $key
+      else
         ssh-keygen -q -N "" -t $alg -b $keysize -C "" -f $key
+      fi
     fi
 
-    cat $message | ssh-keygen -Y sign -n file -f $key > $message.$alg-$keysize-no-options-individual.sig
+    if [ "$alg" == "ssh-dilithium5" ]; then
+      cat $message | ssh-keygen-oqs -Y sign -n file -f $key > $message.$alg-no-options-individual.sig
+    else
+      cat $message | ssh-keygen -Y sign -n file -f $key > $message.$alg-$keysize-no-options-individual.sig
+    fi
 }
 
 create_key_and_sign_ca() {
@@ -32,17 +53,29 @@ create_key_and_sign_ca() {
 
     if [ ! -f "$ca_key" ]; then
         # Create root
-        ssh-keygen -q -N "" -t $ca_alg -b $ca_keysize -C "" -f $ca_key
+        if [ "$ca_alg" == "ssh-dilithium5" ]; then
+          ssh-keygen-oqs -q -N "" -t $ca_alg -C "" -f $ca_key
+        else
+          ssh-keygen -q -N "" -t $ca_alg -b $ca_keysize -C "" -f $ca_key
+        fi
     fi
 
     if [ ! -f "$leaf_key" ]; then
-        # Create leaf
-        ssh-keygen -q -N "" -t $leaf_alg -b $leaf_keysize -C "" -f $leaf_key
-        # Sign the leaf with the root.
-        ssh-keygen -s $ca_key -O clear -I octocat@example.com $leaf_key.pub
+          # Create leaf
+          ssh-keygen -q -N "" -t $leaf_alg -b $leaf_keysize -C "" -f $leaf_key
+          # Sign the leaf with the root.
+          if [ "$ca_alg" == "ssh-dilithium5" ]; then
+            ssh-keygen-oqs -s $ca_key -O clear -I octocat@example.com $leaf_key.pub
+          else
+            ssh-keygen -s $ca_key -O clear -I octocat@example.com $leaf_key.pub
+          fi
     fi
 
-    cat $message | ssh-keygen -Y sign -n file -f $leaf_key-cert.pub > $message.$keypair-leaf-no-options-certificate.sig
+    if [ "$ca_alg" == "ssh-dilithium5" ]; then
+      cat $message | ssh-keygen-oqs -Y sign -n file -f $leaf_key-cert.pub > $message.$keypair-leaf-no-options-certificate.sig
+    else
+      cat $message | ssh-keygen -Y sign -n file -f $leaf_key-cert.pub > $message.$keypair-leaf-no-options-certificate.sig
+    fi
 }
 
 create_key_and_sign_options() {
@@ -63,8 +96,15 @@ create_key_and_sign "ecdsa" 256
 create_key_and_sign "ecdsa" 384
 create_key_and_sign "ecdsa" 521
 create_key_and_sign "ed25519" 256
+create_key_and_sign "ssh-dilithium5" 4864
 create_key_and_sign_ca "rsa" 2048 "ecdsa" 256
 create_key_and_sign_ca "ecdsa" 256 "rsa" 2048
+create_key_and_sign_ca "ssh-dilithium5" 4864 "rsa" 2048
+create_key_and_sign_ca "ssh-dilithium5" 4864 "ecdsa" 256
+create_key_and_sign_ca "ssh-dilithium5" 4864 "ecdsa" 384
+create_key_and_sign_ca "ssh-dilithium5" 4864 "ecdsa" 521
+
+
 
 generate_security_keys=0
 read -p "Generate security key-backed keys (Requires key and user interaction)? [yN] " -n 1 -r
